\name{keboola.r.transformation}
\alias{keboola.r.transformation-package}
\docType{package}
\title{
\packageTitle{keboola.r.transformation}
}
\description{
\packageDescription{keboola.r.transformation}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{keboola.r.transformation}
\packageIndices{keboola.r.transformation}
~~ Runs R transformations ~~
}
\author{
Maintainer: Ondrej Popelka <ondrej.popelka@keboola.com>
}
\keyword{ programming }
\examples{
## TODO
}
