% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keboola.r.transformation.R
\docType{class}
\name{RTransformation-class}
\alias{RTransformation}
\alias{RTransformation-class}
\title{Application which runs KBC transformations in R}
\description{
Application which runs KBC transformations in R
}

